requirements.txt
astutils==0.0.3
crytic-compile==0.1.3
cytoolz==0.10.0
dd==0.5.4
decorator==4.4.0
eth-hash==0.2.0
eth-typing==2.1.0
eth-utils==1.7.0
future==0.17.1
iteration-utilities==0.7.0
manticore==0.3.1
natsort==6.0.0
networkx==2.3
ply==3.11
prettytable==0.7.2
psutil==5.6.3
pydot==1.4.1
pyevmasm==0.2.0
pyparsing==2.4.2
pysha3==1.0.2
PyYAML==5.1.2
rlp==1.1.0
slither-analyzer==0.6.6
TatSu==4.4.0
toolz==0.10.0
wrapt==1.11.2
z3-solver==4.8.5.0
